% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Average_Temperature}
\alias{Average_Temperature}
\title{Average Temperature Data}
\format{
A data frame with 475 rows and 29 variables.
\describe{
\item{Continent}{Continent of the country}
\item{Country}{The name of the country}
\item{City}{The name of the city}
\item{Jan_C}{The average temperature in Celsius in January}
\item{Jan_F}{The average temperature in Fahrenheit in January}
\item{Feb_C}{The average temperature in Celsius in February}
\item{Feb_F}{The average temperature in Fahrenheit in February}
\item{Mar_C}{The average temperature in Celsius in March}
\item{Mar_F}{The average temperature in Fahrenheit in March}
\item{Apr_C}{The average temperature in Celsius in April}
\item{Apr_F}{The average temperature in Fahrenheit in April}
\item{May_C}{The average temperature in Celsius in May}
\item{May_F}{The average temperature in Fahrenheit in May}
\item{Jun_C}{The average temperature in Celsius in June}
\item{Jun_F}{The average temperature in Fahrenheit in June}
\item{Jul_C}{The average temperature in Celsius in July}
\item{Jul_F}{The average temperature in Fahrenheit in July}
\item{Aug_C}{The average temperature in Celsius in August}
\item{Aug_F}{The average temperature in Fahrenheit in August}
\item{Sep_C}{The average temperature in Celsius in September}
\item{Sep_F}{The average temperature in Fahrenheit in September}
\item{Oct_C}{The average temperature in Celsius in October}
\item{Oct_F}{The average temperature in Fahrenheit in October}
\item{Nov_C}{The average temperature in Celsius in November}
\item{Nov_F}{The average temperature in Fahrenheit in November}
\item{Dec_C}{The average temperature in Celsius in December}
\item{Dec_F}{The average temperature in Fahrenheit in December}
\item{Year_C}{The yearly average temperature in Celsius}
\item{Year_F}{The yearly average temperature in Fahrenheit}
}
}
\source{
https://en.wikipedia.org/wiki/List_of_cities_by_average_temperature
}
\usage{
data(Average_Temperature)
}
\description{
List of cities by average temperature in Africa, Asia, Europe,
North America, Oceania and South America (monthly and yearly)
scraped from Wikipedia on March 18, 2024.
The averages of the daily minimum and maximum
temperatures are given. Thus, the actual daytime temperature in a
given month may be considerably higher than the temperature listed
here, depending on how large the difference between daily highs
and lows is.
A real-life data set from Wikipedia
(https://en.wikipedia.org/wiki/List_of_cities_by_average_temperature).
}
\examples{
data(Average_Temperature)
}
\references{
List of cities by average temperature. (2024, March 18).
In Wikipedia. https://en.wikipedia.org/wiki/List_of_cities_by_average_temperature
}
\keyword{datasets}
