% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sunshine_Duration}
\alias{Sunshine_Duration}
\title{Sunshine Duration Data}
\format{
A data frame with 402 rows and 16 variables.
\describe{
\item{Continent}{Continent of the country}
\item{Country}{The name of the country}
\item{City}{The name of the city}
\item{Jan}{Total sunshine hours in January}
\item{Feb}{Total sunshine hours in February}
\item{Mar}{Total sunshine hours in March}
\item{Apr}{Total sunshine hours in April}
\item{May}{Total sunshine hours in May}
\item{Jun}{Total sunshine hours in June}
\item{Jul}{Total sunshine hours in July}
\item{Aug}{Total sunshine hours in August}
\item{Sep}{Total sunshine hours in September}
\item{Oct}{Total sunshine hours in October}
\item{Nov}{Total sunshine hours in November}
\item{Dec}{Total sunshine hours in December}
\item{Year}{Total sunshine hours in a year}
}
}
\source{
https://en.wikipedia.org/wiki/List_of_cities_by_sunshine_duration
}
\usage{
data(Sunshine_Duration)
}
\description{
List of cities by sunshine duration in Africa, Asia, Europe,
North America, Oceania and South America (monthly and yearly)
scraped from Wikipedia on March 18, 2024.
Sunshine duration is a climatological indicator, measuring
duration of sunshine in given period (usually, a day or a year)
for a given location and usually expressed in hours per year, or
in (average) hours per day.
}
\details{
A real-life data set from Wikipedia
(https://en.wikipedia.org/wiki/List_of_cities_by_sunshine_duration).
}
\examples{
data(Sunshine_Duration)
}
\references{
List of cities by sunshine duration. (2024, March 18).
In Wikipedia. https://en.wikipedia.org/wiki/List_of_cities_by_sunshine_duration
}
\keyword{datasets}
