% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Weather_Records}
\alias{Weather_Records}
\title{Weather Records Data}
\format{
A data frame with 270 rows and 6 variables.
\describe{
\item{country_region}{The name of the country or region}
\item{temperature_c}{The extreme temperature (in Celsius) occured}
\item{temperature_f}{The extreme temperature (in Fahrenheit) occured}
\item{town_location}{The name of the location (or town) that the temperature occured}
\item{date}{The date that the extreme temperature occured}
\item{type}{The type of the extreme temperature as cold or hot}
}
}
\source{
https://en.wikipedia.org/wiki/List_of_weather_records
}
\usage{
data(Weather_Records)
}
\description{
List of weather records, i.e. highest (hottest) and lowest (coldest)
temperatures (in Celsius and Fahrenheit)
for each country in the world. The data set is scraped from
Wikipedia on March 18, 2024.
}
\details{
A real-life data set from Wikipedia
(https://en.wikipedia.org/wiki/List_of_weather_records).
}
\examples{
data(Weather_Records)
}
\references{
List of weather records. (2024, March 18).
In Wikipedia. https://en.wikipedia.org/wiki/List_of_weather_records
}
\keyword{datasets}
